+++ src/modules/StoreFrontAdmin/UpdateBanner.js
@@ -2,7 +2,7 @@
 import React, { useState } from 'react'
 import useFetch from "react-fetch-hook";
 
-import { Button, Container, FormControl, Grid, Input, InputLabel, TextField } from '@material-ui/core';
+import { Button, Container, Grid, TextField } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';