+++ src/routes/routes.tsx
@@ -14,6 +14,7 @@ import GnomesScreen from '../modules/altran/screens/GnomesListScreen';
 import GnomeCard from '../modules/altran/screens/GnomeProfileScreen';
 import AdminPanelScreen from '../screens/AdminPanel/AdminPanel';
 import StoreFrontScreen from '../screens/AdminPanel/StoreFront/StoreFront';
+import ReportsScreen from '../screens/AdminPanel/Reports/Reports';
 
 export default function Routes(): ReactElement {
     return (
@@ -27,6 +28,7 @@ export default function Routes(): ReactElement {
                 {/* <Route path="/airbnb" component={Airbnb} /> */}
                 <Route path="/altran/user/:id" component={GnomeCard} />
                 <Route path="/altran" component={GnomesScreen} />
+                <Route path="/admin-panel/reports" component={ReportsScreen} />
                 <Route path="/admin-panel/store-front" component={StoreFrontScreen} />
                 <Route path="/admin-panel" component={AdminPanelScreen} />
                 <Route path="/" component={Home} />