+++ src/screens/AdminPanel/Reports/Reports.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import Footer from '../../../components/elements/Footer/Footer';
+import Header from '../../../components/elements/Header/Header'
+
+import Copyright from '../../../components/elements/Copyright/Copyright';
+
+import Dashboard from '../../../modules/Reports/Dashborad';
+
+export default function StoreFront() {
+
+    return (
+        <>
+            <Dashboard />
+            <Copyright />
+        </>
+    );
+}